le patch
added manpages for notes commands, added manpages for all the individual theme commands,

fixing an issue involving exporting themes. exported themes are now going to just the stringified JSON of the theme, because it's just better.

I also am making a function where it checks if a JSON object has a specific structure for any other import/exports because yes.